+++ includes/cleanup-frontend/cleanup-frontend.php
@@ -1,0 +1,7 @@
+<?php 
+
+/* Disable XML-RPC and Pingback */
+require_once( __DIR__ . '/xmlrpc-and-pingback.php' ); 
+
+/* Clean <head>-section */
+require_once( __DIR__ . '/head.php' );