+++ includes/custom-scripts/custom-scripts.php
@@ -1,0 +1,7 @@
+<?php 
+
+/* Allow admin to add scripts to entire site */
+require_once( __DIR__ . '/site-scripts.php' );
+
+/* Allow admin to add scripts to specific posts */
+require_once( __DIR__ . '/post-scripts.php' );