+++ includes/js/admin-social-links.js
@@ -1,9 +1,0 @@
-jQuery(document).ready(function($) {
-
-    //* Move Social Link
-    $( ".ejo-social-links tbody" ).sortable({
-        revert: true,
-        handle: ".move-item",
-    });
-
-});