+++ includes/social-media/social-media.php
@@ -1,0 +1,4 @@
+<?php
+
+/* Social Media Links */
+require_once( __DIR__ . '/links/social-media-links.php' );