+++ includes/widgets/widgets.php
@@ -1,0 +1,10 @@
+<?php 
+
+/* Manage default WordPress Widgets */
+require_once( __DIR__ . '/manage-default-widgets.php' );
+
+/* Allow widget templates to be loaded from theme directory */
+require_once( __DIR__ . '/widget-theme-templates/widget-theme-templates.php' );
+
+/* Footer Line Widget */
+require_once( __DIR__ . '/footer-line-widget/footer-line-widget.php');